+++ Core/Libraries/Source/WWVegas/WW3D2/Backend/DX8Backend.h
@@ -46,8 +46,8 @@ public:
 
     virtual void Set_Vertex_Buffer(const VertexBufferClass * vb, unsigned int stream) override;
     virtual void Set_Vertex_Buffer(const DynamicVBAccessClass & vba) override;
-    virtual void Set_Index_Buffer(const IndexBufferClass * ib, unsigned short index_base_offset) override;
-    virtual void Set_Index_Buffer(const DynamicIBAccessClass & iba, unsigned short index_base_offset) override;
+    virtual void Set_Index_Buffer(const IndexBufferClass * ib, unsigned int index_base_offset) override;
+    virtual void Set_Index_Buffer(const DynamicIBAccessClass & iba, unsigned int index_base_offset) override;
     virtual void Set_Index_Buffer_Index_Offset(unsigned int offset) override;
 
     virtual void Set_Shader(const ShaderClass & shader) override;
@@ -75,19 +75,19 @@ public:
     virtual void Set_Light_Environment(LightEnvironmentClass * light_env) override;
     virtual LightEnvironmentClass * Get_Light_Environment() const override;
 
-    virtual void Draw_Triangles(unsigned short start_index,
-                                unsigned short polygon_count,
-                                unsigned short min_vertex_index,
-                                unsigned short vertex_count) override;
+    virtual void Draw_Triangles(unsigned int start_index,
+                                unsigned int polygon_count,
+                                unsigned int min_vertex_index,
+                                unsigned int vertex_count) override;
     virtual void Draw_Triangles(unsigned int buffer_type,
-                                unsigned short start_index,
-                                unsigned short polygon_count,
-                                unsigned short min_vertex_index,
-                                unsigned short vertex_count) override;
-    virtual void Draw_Strip(unsigned short start_index,
-                            unsigned short index_count,
-                            unsigned short min_vertex_index,
-                            unsigned short vertex_count) override;
+                                unsigned int start_index,
+                                unsigned int polygon_count,
+                                unsigned int min_vertex_index,
+                                unsigned int vertex_count) override;
+    virtual void Draw_Strip(unsigned int start_index,
+                            unsigned int index_count,
+                            unsigned int min_vertex_index,
+                            unsigned int vertex_count) override;
 
     virtual void Set_Vertex_Shader(unsigned int vertex_shader) override;
     virtual void Set_Pixel_Shader(unsigned int pixel_shader) override;
